Reading from init file
Updated game and some related classes to init the world map from a
external file, rather than from within the game class. It needs
polished, but has many checks to avoid an unknown death by syntax. Also
added a mapViewer to see a list of the rooms with their items and exits
quickly.